+++ rune/src/lint/rules.zig
@@ -103,6 +103,7 @@ const RULES = [_]RuleEntry{
     .{ .rule = .unique_index_redundant_with_pk, .handler = index_rules.checkUniqueIndexRedundantWithPk },
     .{ .rule = .unique_index_redundant_with_unique, .handler = index_rules.checkUniqueIndexRedundantWithUnique },
     .{ .rule = .unindexable_type_indexed, .handler = index_rules.checkUnindexableTypeIndexed },
+    .{ .rule = .unsigned_overflow_risk, .handler = portability_rules.checkUnsignedOverflowRisk },
     .{ .rule = .table_no_index, .handler = index_rules.checkTableNoIndex },
     .{ .rule = .index_name_too_long, .handler = index_rules.checkIndexNameTooLong },
     // View validation rules (moved to view.zig)